Tsylia / Disable C(R)UD and Block operations at the blocked providers (#1202)
* The blocked provider (director) has the possibility (besides the blocked deputy provider and the blocked provider admin ) to log in to the system.

* Forbidden the CRUD (w/o R) and Block operations for the blocked provider with workshops, parents, applications, and providers admins (also deputies).

* Fix some tests.

* Added the blocking the provider admins and deputies after the blocking the provider.

* Added BlockingType model (manually or automatically deleting).
Updated ProviderAdmin model.
Changed ProviderAdminService.
Changed AuthController. Blocked provider, deputy and provideradmin can log in.

* Forbidden the CRUD (w/o R) and Block operations at the blocked provider.

* Add migration

* Fixed code smells.

* Fixed bug

* Added TODO to the tests at ProviderControlleTests.

* Code review comments.

+++ OutOfSchool/OutOfSchool.WebApi/Controllers/V1/ProviderAdminController.cs
@@ -6,6 +6,8 @@ using OutOfSchool.Services.Enums;
 using OutOfSchool.Services.Models;
 using OutOfSchool.WebApi.Common;
 using OutOfSchool.WebApi.Models;
+using OutOfSchool.WebApi.Models.Application;
+using OutOfSchool.WebApi.Services;
 
 namespace OutOfSchool.WebApi.Controllers;
 
@@ -16,15 +18,21 @@ namespace OutOfSchool.WebApi.Controllers;
 public class ProviderAdminController : Controller
 {
     private readonly IProviderAdminService providerAdminService;
+    private readonly IUserService userService;
+    private readonly IProviderService providerService;
     private readonly ILogger<ProviderAdminController> logger;
     private string path;
     private string userId;
 
     public ProviderAdminController(
         IProviderAdminService providerAdminService,
+        IUserService userService,
+        IProviderService providerService,
         ILogger<ProviderAdminController> logger)
     {
         this.providerAdminService = providerAdminService;
+        this.userService = userService;
+        this.providerService = providerService;
         this.logger = logger;
     }
 
@@ -49,6 +57,21 @@ public class ProviderAdminController : Controller
     public async Task<ActionResult> Create(CreateProviderAdminDto providerAdmin)
     {
         logger.LogDebug($"{path} started. User(id): {userId}.");
+
+        if (providerAdmin == null)
+        {
+            return BadRequest("ProviderAdmin is null.");
+        }
+
+        if (await IsProviderBlocked(providerAdmin.ProviderId).ConfigureAwait(false))
+        {
+            return StatusCode(403, "Forbidden to create the provider admin at the blocked provider");
+        }
+
+        if (await IsCurrentUserBlocked())
+        {
+            return StatusCode(403, "Forbidden to create the provider admin by the blocked provider.");
+        }
 
         if (!ModelState.IsValid)
         {
@@ -86,6 +109,21 @@ public class ProviderAdminController : Controller
     [HttpPut]
     public async Task<IActionResult> Update(Guid providerId, UpdateProviderAdminDto providerAdminModel)
     {
+        if (providerAdminModel == null)
+        {
+            return BadRequest("ProviderAdmin is null.");
+        }
+
+        if (await IsProviderBlocked(providerId).ConfigureAwait(false))
+        {
+            return StatusCode(403, "Forbidden to update the provider admin at the blocked provider");
+        }
+
+        if (await IsCurrentUserBlocked())
+        {
+            return StatusCode(403, "Forbidden to update the provider admin by the blocked provider.");
+        }
+
         if (!ModelState.IsValid)
         {
             return BadRequest(ModelState);
@@ -133,6 +171,16 @@ public class ProviderAdminController : Controller
     {
         logger.LogDebug($"{path} started. User(id): {userId}.");
 
+        if (await IsProviderBlocked(providerId).ConfigureAwait(false))
+        {
+            return StatusCode(403, "Forbidden to delete the provider admin at the blocked provider");
+        }
+
+        if (await IsCurrentUserBlocked())
+        {
+            return StatusCode(403, "Forbidden to delete the provider admin by the blocked provider.");
+        }
+
         var response = await providerAdminService.DeleteProviderAdminAsync(
                 providerAdminId,
                 userId,
@@ -164,6 +212,16 @@ public class ProviderAdminController : Controller
         {
             logger.LogDebug("IsBlocked parameter is not specified");
             return BadRequest("IsBlocked parameter is required");
+        }
+
+        if (await IsProviderBlocked(providerId).ConfigureAwait(false))
+        {
+            return StatusCode(403, "Forbidden to block the provider admin at the blocked provider");
+        }
+
+        if (await IsCurrentUserBlocked())
+        {
+            return StatusCode(403, "Forbidden to block the provider admin by the blocked provider.");
         }
 
         var response = await providerAdminService.BlockProviderAdminAsync(
@@ -293,6 +351,11 @@ public class ProviderAdminController : Controller
     {
         logger.LogDebug($"{path} started. User(id): {userId}.");
 
+        if (await IsCurrentUserBlocked())
+        {
+            return StatusCode(403, "Forbidden to reinvite the provider admin by the blocked provider.");
+        }
+
         var response = await providerAdminService.ReinviteProviderAdminAsync(
                 providerAdminId,
                 userId,
@@ -313,4 +376,11 @@ public class ProviderAdminController : Controller
                 return Ok();
             });
     }
+
+    private async Task<bool> IsCurrentUserBlocked() =>
+        await userService.IsBlocked(userId);
+
+    private async Task<bool> IsProviderBlocked(Guid providerId) =>
+        await providerService.IsBlocked(providerId).ConfigureAwait(false);
+
 }
